kafka-paperplane-sink/requirements.txt
python-decouple
honcho
yajl
-e git+https://github.com/wsacin/kafka-paperplane.git#egg=kafka-paperplane
